requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu129
numpy>=1.26.4
gdown>=3.10.1
Pillow>=8.0.0
requests>=2.25.0
tqdm>=4.62.0
matplotlib>=3.5.0
torch==2.8.0+cu129
torchvision==0.23.0+cu129
scikit-image>=0.18.0
scikit-learn>=1.0.0
pandas>=1.3.0
flask>=2.0.0
flask_cors>=4.0.1
mtcnn-pytorch>=0.1.0
gunicorn
retinaface-pytorch
insightface
onnxruntime-gpu
mediapipe
dlib
ultralytics
facenet-pytorch>=2.5.3
typing-extensions
pydantic
albumentations
opencv-contrib-python>=4.5.5.64
fire>=0.4.0
resnet